Check if sender is part of the org
If PR sender is a public member (because we can't otherwise with github
app to have that info) on the organization where the pr has been sent.

If user that submitted the PR is owner of the repo then allow it too.

There is going to be a bit more logic here for OWNERS but let not go too
ahead of ourselves.

partial addressing #39

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>