Add event_type in created pod
Refactor code to use github.ParsePayload to parse the event type.
This is the premice works to get proper separation of all event type.

We address a minor hijack probability where a rogued user would be able
to use the recheck event to DDOS a CI.

Triggers is stil buggy with a full payload, we probably will need to change the
technique of passing information in there unless we start to redo everything as
a Webhook interceptor and just not deal with triggers anymore :\