Match pipelinerun by annotation
Introduce pipeline annotation based event matching.

If we have multiple pipelineruns in the .tekton directory, we go over the
annotations and match the event on the pipeline annotation.

For example :

metadata:
 name: pipeline-pr-main
 annotations:
    pipelinesascode.tekton.dev/on-target-branch: "[main]"
    pipelinesascode.tekton.dev/on-event: "[pull_request]"

will match the pipelinerun pipeline-pr-main if the github event target the main
branch and is coming from a pull_request.

If there is multiple pipeline matching the same event, it will match the first
one.

TODO: We currently enforce annotations to be in PipelineRun, maybe this can be
optional and by default if there isn't annotation the target branch will be the
repository default branch (ie: main) and the event will be pull_request.