Allow more flexibility for branch matching
We were simply matching the event branch name to whatever we have in
Repository.

In push this would come up as `refs/heads/main` and on pull as `main`
which would be confusing.

Now the user can simply specify `main` even tho we receive a
refs/heads/main and this will match it.

Added glob support too with the github.com/gobwas/glob library.
If we get something like refs/tags/1.* it will match every tags start
with "1."

Refactored all the testbranch test in a testsuite.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>