Fix friendly matching on branch
We were not doing it properly on matching the branch.

There is multiple place where we need to do that "friendly matching"
when we get the payload matching it to the pipelinerun and to the
Repository information.

We were only doing on repository information and not on pipelinerun.

Refactored all of thise in the same go package so it makes it easier
in future and improved all the tests.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>