Handle ui rerequest on push
We were only handling it on pull_request, we had to do a whole different
thing to make it  work on push.

We set the sender to the one who clicks on the rerequest on push,
compared on pull_request which we assume the sender is pr author.

The only way to be able to do a rerequest on push is to be a owner of
the repository, so that should be harmless.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>