Add E2E tests for PullRequest, Push, Retests
Add plumbing and some tests for PR, Push, Rerequest and Retest

The tests assume there is a cluster with pipeline as code setup. There
is a test repo available with a URL and TOKEN set.

It will be flexible enough to test multiple scenarios and finally have
e2e tests here.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>