Remove by default question about repo crd
If we can guess the resource name for the repo crd then just use it
don't ask for it. We don't care about conflict yet but if it's become a
issue maybe we will check for it and add a numbered suffix.

Use can still override on the command line if they wanted too.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>