Add CloneURL, Lower Secrets name and PRName to status
Add a CloneURL to info.Events for Bitbucket Server mostly but may be use by
others.

Lower secrets name, since an org is a project in bitbucket server and a project
can be uppercase which bugs out the secret creation.

Add PipelineRunName to status for VCS drivers to add, which in case of bitbucket
server is useful.