Cleanup by the actual PR name in pr file
There was a bug since we moved to repo without filtering target where we
were erasing every PR attached to a repo, which mean push and pull and
everything.

We now filter by the prName from the pipelinerun as defined in .tekton/
dir, which would even make it more scalable to the additional filtering
we may want to add in the future (I am looking at you path filtering).

Fixes bug #339

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>