feat: add pipelinerunname to the github check
We add the pipelinrunname on github. This probably be more useful whenever we
implement multi pipeline match on events but this gives a nice way to
see the pipeline that has been tested.

maybe improvement in the future to allow via an annotation to customize
the label output instead of just the PR name.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>