Document resolve behaviour between cli and live
There is a difference in behaviour between cli and running on CI.

On live, we grab everything in .tekton and resolve it even if it's not specified
in annotations. Added an explicite test for it.

On CLI the user need to explicitely specify the templates/dir to grab
from.

Document it and rework a bit the documentation.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>