fix getting repo spec credentials with no api url
When we run in bitbucket cloud we don't need to set a api url and
documented as is. Don't bail out on one. The other providers who need an
api url will anyway check for it in SetClient and bail out if not set
properly.

This was introduced as a side effect when running the tests since we
were constructing the tests with GitProvider struct set and we check in
the code when it's not nil. Fix the test fixture to not passing the
strucs unless we have a URL or user passed.

Closes #416

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>