Try to detect if pac is installed via operator
When installing via openshift-pipelines "downstream" operator. PAC would
get installed in openshift-pipelines.

Strategy to detection is then :

if user has a -n ns flag, try to see if pac is installed in ns
try to detect pac in openshift-pipelines ns (downstream)
try to detect pac in pipelines-as-code ns  (upstream)

The way to detect it is to check the configmap installed in the
openshift-pipelines namespace. This makes this detection compatible with
old way (via taskrun) and the new way via controller.

As far goes for for permissions usually the folks who tries to do
bootstrap will have to have anyway right access to that ns to be able to
bootstrap so we should be good.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>