Repo create cmd: convert username to lowercase and then use it
if a user has capital letters in username then repo create cmd
returns an error:
```
Error: Repository.pipelinesascode.tekton.dev "Abc-pac-test" is invalid:
metadata.name: Invalid value: "Abc-pac-test": a lowercase RFC 1123 subdomain must consist of
lower case alphanumeric characters, '-' or '.', and must start and end with an alphanumeric character (e.g. 'example.com',
regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*')
```
this fixes by converting to lowercase and then use it.

Signed-off-by: Shivam Mukhade <smukhade@redhat.com>