[gh app] Update exiting check run if there is one
When there was an already exiting check_run we discover it. It fixes the
issue when we :

- Create a commit
- There is a failure before we have checked out the pipelinerun match
  (ie: hub is down and fail to get the taskRef remotely)
- We do a /retest and the error is fixed.
- We would still have the status of the first failure as well a second
  successfull one.

We now list the check_runs attached to the ref created by our
ApplicationID and re-use that ID instead of creating a new one.

There is a slight bug in github interface, if we update the check run
and set as pending, it doesn't reflect it in the UI. It does change it
when the check_run went from failure to successfull so that's fine.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>