Use defaults for pipelinerun template in repo create cmd
this updates the repo create cmd to skip asking event type to user
and use default '[pull_request, push]' as event type for template
generation. the focus of repo create cmd is to create repo and
generation of template is an additional feature so we use defaults
for that now and skip asking user for input.

Signed-off-by: Shivam Mukhade <smukhade@redhat.com>