fixes comment regex to allow string along with /retest and /ok-to-test
comments with some string and (retest)/(ok-to-test) were failing as regex
was failing to parse it, this update the regex.
(/retest and /ok-to-test must be on a separate line with nothing else on
that line, then the comment will be processed)
now, following comments are allowed
valid eg.
```
abcde
/retest
xyz
```

```
/lgtm
/ok-to-test
```

invalid eg.

```
/retest abc
```
or
```
/ok-to-test xyz
```

Signed-off-by: Shivam Mukhade <smukhade@redhat.com>