Fix watching for pipelinerun timeout
First we define a ridiculous large number for taskrun timeout, so the
pac taskrun doesn't timeout before the user defined one.

Second we are making sure we are not using the timeout configuration
instead of the constant (which was 10minutes by default)

Third we are not just warning anymore but actually exit.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>