match multiple pipelineruns on the same events
we now gets all the matched condition from the matcher in the same struct

we did a bunch of refactoring to break out .Run which got moved to its own

Make event able to process multiple pipelineruns

We run a goroutine in a workgroups which we wait for. This will be run in the
main goroutine which started from the sinker.

We disable secret auto creation for now when running multiples, it has many
issues we need to fix regardless.

We retries if we can't update the RepoStatus until we are able to do it.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>