Try to import the same fix as what we have on push
On push pipeline we did a weird fix which make no sense but works.

https://github.com/openshift-pipelines/pipelines-as-code/commit/92de7b4f2d0dfae5afbbf013068c3d4600e08c07

It was failing before because for whatever reasons it could not find the
shell scripts in hack  :

```
[release] /tekton/scripts/script-0-vkrmt: line 4: hack/generate-releaseyaml.sh: No such file or directory
```

while the PATH are set properly and everything,

doesn't make sense at all but with this "fix" it was working :

```
[upload-release-yaml : release] TAG SHA: 92de7b4f2d0dfae5afbbf013068c3d4600e08c07
[upload-release-yaml : release] Create or update branch: nightly
[upload-release-yaml : release] refs/heads/nightly has been updated to 92de7b4f2d0dfae5afbbf013068c3d4600e08c07
[upload-release-yaml : release] Upload file release.k8s.yaml to release.k8s.yaml based on 92de7b4f2d0dfae5afbbf013068c3d4600e08c07
[upload-release-yaml : release] Upload file release.yaml to release.yaml based on 34f810968d2a4ee5deb79b9feabadcc7b95cae9a
```

so we import that same "fix" here on release pipeline since those are
harder to debug we just pray that the "fix" fixes it

again i have no idea what i am doing!

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>