add provider type to the labels
we add a label with the provider type.
we don't put it in the repostatus since it doesn't make sense.
I wanted to update the cli/describe to show it, but that may be more
difficult since we rely on repostatus for it. may need a new struct and
new trickery

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>