Add repo-url annotation to generated PipelineRun
Add the repo url to get the real repo URL where it's coming from.  We could do
this as well with event.Organization and event.Repository but since those are
labels they get transformed to please lord k8s and most of the time that's fine
but with gitlab you can have / in "Organization" (which is actually called
Project there) and since they / gets transformed to - by pkg/formatting/k8labels
we would not know the proper name.

We actually do need the proper name since we would not know where it's coming
from to update status with gitlab.

so now the reconciler/watcher will use that value to fill up the events.org and
events.repo. Added bunch of test to handle all use case i can think of.

Acked-by: Chmouel Boudjnah <chmouel@redhat.com>