Enable TestGiteaBadYaml
This test goes deep, it would use a bad yaml that has bad format and
check inside the controller logs (it has a random unique id as the
pipeline name) to make sure we errored out properly on it.

More stuff to come with regard to dig into controller logs (or events
log namespace when this will be imp) tests.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>