Fix TestGiteaConfigMaxKeepRun race
We had a race in TestGiteaConfigMaxKeepRun since we were not waiting
that the repo status was updated to list the number of prs. It was
working sometime on k8 but fails on openshift since it's a slower api
server.

We wait for the repo status to be updated now before we list the number
of pr running..

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>