Reuses skipped checkrun, as we can't remove it
skipped runs were not getting removed a authorised user adds /ok-to-test
on a pull req from a new user, this is to fix that case.

we cannot delete a checkrun so we reuse it for the pipelinerun, when the
CI is triggered later.
as we execute pipelineruns parallely, to make sure only one pipelinerun uses that
checkrunid we use mutex to store that checkrunid.