Fixes target branch match, if annotation has refs/head
in case of pullreq, base branch is just a name without
refs/heads but if user in annotation has provided values in regex
form eg. refs/head/rel-* then it use to fail as we didn't have
refs/heads in base branch prefix, for those cases we add
prefix and check if they match.

Signed-off-by: Shivam Mukhade <smukhade@redhat.com>