Workaround some gitea bugs
Remove the <br> cause gitea don't like anymore and we can use \n

Add a random number (to be removed) to the comment because gitea 422
when it's exact same string, that only happen when using the API not
sure why it doesn't show on UI....

Add loop for wait for events to come up, we could not just sleep
blindly, because it may be slow and then fails. 😴

Fix e2e tests to properly run on different gitea host than the local
one.

Skip TestGiteaConcurrencyExclusivenessMultipleRuns test (just can't
figure out)

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>