Fix tkn pac resolve with remote task
We were not getting the pac settings  properly to get the default hub
catalog name and it would fail.

Don't output log.Info on resolve, maybe we can add a flag in the feature
to enable it.

Plug into the e2e test as well that would detect this kind of defect.

Fixes #1039

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>