Fix checking token expiration on webhook
We were inverted the logic so it was wrong.
We would not be able to reproduce it when the header was not set since
the header was not set on unlimited life token and other cases i am not
totally sure about.

Add more test to handle all use cases

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>