Fix tkn pac resolve
I can't figure out why the apiVersion and Kind is not getting populated
on CLI while working when generated from the controller.

let's add it manually which seems to generate proper PipelineRun

We do have e2e tests covering all of this code path but it seems to
occur only from a binary and not when launched from go code, so i guess
we can't do much about it.

Fixes #1178

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>