Report pipelinerun error properly
We didn't previously reported the startPR erros onto the user interface.
We now do it properly which would give an indication to the user why the
pipelinerun wasn't started.

Could not figure out how to fake a validation error from tekton
controller in unittest, but the added e2e covers it for the gitea comment
interface and on the controller.

Improved the error reports all over this function, to let know exactly
what action caused this error.

Fixes #1224
https://issues.redhat.com/browse/SRVKP-2992

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>