Add information on 'tkn pac info' command for Pipelines as Code installation
This commit adds the 'tkn pac info' command, which provides useful
information about the installation of Pipelines as Code. The command
displays the location and version of the installation, as well as an
overview of all the Repositories CR created on the cluster and their
associated URLs. It also includes details of any installed GitHub Apps.

https://issues.redhat.com/browse/SRVKP-2977
Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>