Add setting to only trusts users from repo
By default, when using for example the GitHub provider, if your
repository is belong to an organization, the users belonging to that
organization are granted automatic permissions to initiate the
pipelines. However, considering the potential existence of malicious
users within a large organization and the need to exercise caution,
there may be certain repositories for which we do not wish to extend
trust to everyone.

To address this, you have the option to deactivate this functionality by
configuring the repository settings and setting the
`only_trusts_users_from_repository` parameter to `true`.

example:
```yaml
spec:
  settings:
    only_trusts_users_from_repository: true
```

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>