Add --v1beta1 flag for pipelinerun output
On certain clusters, the conversion from v1beta1 to v1 in Tekton may not
function correctly, leading to errors when applying the resolved PipelineRun on
a different cluster that doesn't have the bundle feature enabled. To resolve
this issue, you can use the `--v1beta1` flag (or `-B` for short) to explicitly
output the PipelineRun as v1beta1 and work around the error.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>