Allow setting policies on actions
Policies is a new concept which let you allowing teams to do some
actions and reject member who are not part of those teams.

Current actions supported are `pull_request` and `ok_to_test` (more to
come in the future)

See the documentation attached to this PullRequest for more description
on how to use this feature.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>