Optimize matching pipelineruns with remote
We try to optimize the matching of the pipelineruns with the remote tasks when
we have a large set of pipelineruns in the .tekton/ directory. We don't want to
fetch all the remote tasks every time just for the matching. We first try to
match without fetching the remote tasks and if we have a match we try to
match with fetching the remote tasks.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>