GH App, make sure token generated works for itself
When we were generating a token we would do this for the installation
token where the repo is installed.  However, some installations would
not have the repo where it comes from (or target on incoming webhook)
and due fail.

Fixed the tests along the way which wasn't tested thing fully and made
them more robust.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>