Only create prun when a MR update with push happen
We were creating a pipelinerun when an MR was updated for every event,
which mean every actions like adding labels or setting a status will
create a new PipelineRun.

We only want it to get it kicked when it is created or reopen or when a
there is a commit push attached to that update.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>