Don't create pipelinerun when an MR update
We were creating a pipelinerun when an MR was updated, which mean every
actions like adding labels or setting a status will create a new PipelineRun.

We only want it to get it kicked when it is created or reopen! or when a
push to the base branch is done to a PR.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>