Fallback to other ACL rules if the policy is set
On failure we were not falling back to other ACL rules and we were not
checking if the other ACL rules were matching (like OWNERS file).

Tests is more comprehensive and check that we are falling back properly
and as well that the target is matching the OWNERS file.

We now properly dump the errors/allowance in the kube events streams.

Try to make sharing functions between gitea and GitHub as much as
possible.

Fix https://issues.redhat.com/browse/SRVKP-3542

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>