Fix using gitclone clustertask when detected avail
We now detect cluster task by default (unless user set it to false) for
the git-clone if available we use it. If not we use the hub task.

We cannot use templates in here because the console consumes those
directly (which sucks).

We could have a post job generating all templates in release/nightly
like we do for release.yaml and have the console uses it maybe but
that's for another day.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>