Fix secret cleanup issue
When max-keep-runs is exceeded, The GitHub basic auth secret
for the latest pipelineRun is deleted.

For example, if max-keep-runs is set to 1, kick off a pipelineRun.
when it's done, kick off another instance of the same pipelineRun.
When the second pipelineRun is done and pac performs a cleanup
to delete the first pipelineRun, the deletion is proper for first
pipelinerun, but it also deletes the secret associated
with the second pipeline run

Signed-off-by: Savita Ashture <sashture@redhat.com>