Fix policy settings
Policy setting was broken, this would always get the other ACLs check to
run even when the user was not there.

The tests was also not testing the policy setting properly.

The E2E tests is hard to test on GitHub since we need to create another
two others users and another organisation with a bunch of teams to test
it properly. Hopefully if we can get our own GitHub enterprise instance
we would be able to do those in the future.

I have done those tests manually:

An organisation with three users:

chmouel
sam
rubben

A team with the users sam and chmouel in there but not rubben.

I have created a repo CR with the policy settings like this:

```yaml
  settings:
    policy:
      pull_request:
      - pull_request
```

the user sam gets allowed to run CI but the user rubben get disallowed.

If I issue a comment with /ok-to-test from the user sam on the rubben PR
it will be disallowed (since it only to run the pr and not issue
ok-to-test).

I changed the policy on the repo settings with the ok-to-test section
allowing the group pull_request:

```yaml
  settings:
    policy:
      ok_to_test:
      - pull_request
      pull_request:
      - pull_request
```

and then I can issue /ok-to-test from the user sam on the user rubben
and it will be tested.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>