Add private URL example to GitLab URL
Atm we do have in a note that private instances (i.e. self-hosted
enterprise/private instances) of GitLab are not automatically detected.

However we could make it easier for the developers to grasp what do we
mean, by adding said field (even in a commented out fashion) in the
example

This commit add the url for example and as well make sure we use the
word GitLab instead of gitlab by enforcing a rule with vale.

https://issues.redhat.com/browse/SRVKP-3735