Make sure to sort secrets by longest when hiding
When we hide secrets in the logs we need to make sure that we start from
the longest one, otherwise if we have two secrets with the same prefix the
shortest one may hide first and leak the end of the longest one. Since
we are in a loop and text is replace the substitutions would not occurs.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>