Multi namespace support
This introduce the concept of multiple GitHub application from the same
PaC installations.

You will need to deploy another controller with another set of
environment variable pointing to a new secret and a ConfigMap to
configure this, see the documentation showing how to install this.

e2e tests now tests concurrently (for some tests) on GHE instance
(ghe.pipelinesascode.com) and public GitHub.

The watcher is unique it read the information from where the
controller/secret/configmap the request come from on the PipelineRun
annotation and resolves it to the right place with the right secret to
the right place.

We now stores as well some of of the controller info in the context
instead of the global variables, this involves a large refactoring
around the code.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>

add openshift routes for second-controller.py

Add support for output openshift routes for second-controller.py