Revised if statement to address validation of names
Previously, the code was checking a slice object,
[]*tektonv1.PipelineRun, with nil.
This caused an issue where if MetadataResolve function
returned an error and an empty slice ([]),
the code wouldn't enter the error condition.
However, the subsequent part of the code couldn't
process the empty slice, resulting in an error
in the PAC controller stating
cannot match pipeline from payload to a
pipelinerun in .tekton/ dir.
This lack of clarity in the error message made
debugging challenging.

The correction involved checking the
length of the object to address this issue.

Signed-off-by: Savita Ashture sashture@redhat.com