Resolve inconsistent status of a PipelineRun after a failure
In Pipelines as Code for some CI failures failed checkrun persist
like below

Pipelines a Code CI - Failed

on the GitHub status page.

These are the 2 scenarios

1. If there is a pull request on which the CI has failed checkrun,
updating the PR results in the override of the failed check run.
This is because a PR update generates a new commit SHA,
leading to the creation of new check runs for the updated SHA.

2. If there is a pull request on which the CI has failed checkrun,
if a user adds a /test or /retest comment on the same commit SHA
without updating the PR, the failed check run will be retained.
Alongside this, new check runs will be added for the existing SHA.

However, in both scenarios, a challenge arises when there
are multiple pipeline runs (n pipelineruns) in the .tekton
directory, resulting in the creation of
n check run IDs for each failed pipeline run.

With the proposed fix, Pipelines as Code now effectively
addresses this issue.
The solution ensures that, for failed pipeline runs,
the check run is appropriately overridden without
generating n check run IDs for each failed pipelinerun.