Fix release pipeline for wrong version
During release, pipeline is pushing the branch as release-0.24.0
instead of release-v0.24.0 and similary same version is getting replaced
in release yaml and also version file

This is to fix that behaviour. We are overiding the version variable
in the for loop which is defined before the for loop containing
the value of version getting released, and it is getting overwritten as the
for loop is running in same shell. Declare a new name for the variable
inside the for loop so that it does not override the original variable
which is later getting used in script