Add on-comment annotation to match PipelineRun
We now support the `on-comment` annotation to match a PipelineRuns on
pull requests.

This is only supported on pull_requests and on GitHub, Gitea and GitLab
providers.

Additionaly we have added a trigger_comment to the standard parameters
for pipelinerun to use.

This allows getting the comment that triggered the pipelinerun when
using a GitOps command or a custom GitOps command via the on-comment
annotation.

This will allow some powerful workflow like a PipelineRun that does:

/merge branch

and user can then have a PipelineRun that will trigger on the `/merge
.*` regexp and then parse the output and merge the PR into `branch`
without having to do many API calls to figure out which comments
triggered the PipelineRun.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>