gitlab: Fix posting prun status on gitops comments
we were not posting the status on gitops comments, this is because we
were not catching the Note events as defined by the gitlab provider.

better functional tests to check the actual comments has been posted
properly on gitops comments for gitea.

This is for the release-v0.24.x branch since we have generics function
for this in the main branch.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>