Gitlab: Fix posting prun status on GitOPS comments
We were not posting the status on GitOPS comments, this is because we
were not catching the Note events as defined by the Gitlab provider.

Better functional tests to check the actual comments has been posted
properly on GitOPS comments for gitea.

This has been backported as well for release-v0.24.x in
https://github.com/openshift-pipelines/pipelines-as-code/pull/1585

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>