e2e: separate cancel and retest GitOPS comments test
We were mixing the retest and cancel GitOPS comments in the same test
with one test function assuming the result of the previous one.  This
is not a good practice and we should separate them to make it easy to
debug and understand.

The tests was wrong as well, since we were checking for 3 success status
in conditions but since we are canceling the pipeline we don't get
success. So now checking properly the PipelineRun is cancelled properly.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>