Remove detection of old basic auth secret
I have added that check in 42ff2920ea5bb7b6db37834314aa49bdf69b4011
(April 2022) when we migrated from one secret name to another.

We can safely remove it now but keep that function for future checks if
needed when doing a similar migration.

Improve documentation for the developers on how to add new checks.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>