fix race conditions on pr concurrency
when storing a pointer in the context, the pointer get nil sometime on
high load which causes some of the pipelinerun issue

we fix this by passing the value around instead of passing the pointer
in context. There is more going on in this bug that meet the eyes but
that seems to fix it.

remove all the StoreInfo and GetInfo functions and use the value directly.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>