Don't try to cleanup Pending PipelineRuns
When using max-keep-runs annotations with concurrency, the cleanups
process on the watcher was trying to cleanup Pending PipelineRuns. This
would cause the watcher to get stuck and not able to process its queue.

We are now avoiding it the same way we do for Running PipelineRuns.

Improved TestGithubSecondPullRequestConcurrencyMultiplePR to include
max-keep-runs and /retest over concurrency

Increase the timeout for running the e2e tests since that test is
getting a bit slower.

Jira: https://issues.redhat.com/browse/SRVKP-4266
Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>