Only cleanups PRs after reconciliation
I experienced some issues where the watcher was getting stuck when there
is a race between the deletion of the PR and the reconciliation.

The high load would make it slow to know which PR to get deleted, and
the reconciliation would be stuck trying to reconcile on a deleted PR.

Stress tested this over 150 concurrent PR of 3 PipelineRuns with a
concurrency of 1 and a max-keep-runs of 1 for each pipelinerun and the Q
processed properly until the end all PipelineRuns without getting stuck
(like it was before the change).