Make sure we get the oldest repository
When we have multiple matches for the same repository, we need to make
sure we only get the oldest one. This is important since we pac assumes
the first one is the one that should be used, to avoid an hijacking
scenario.

This should never happen in practice since we have our webhook that make
sure we only have one repository per URL per cluster. If for whichever
reasons the webhook is crashing or the admin has disabled it then at
least we are covered

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>