add golangci-lint errcheck warning
we were getting this warning:

```
WARN [config_reader] The configuration option `linters.errcheck.exclude`
is deprecated, please use `linters.errcheck.exclude-functions`.
```

so removing .errcheck.txt and adding the functions to the .golangci.yml

add a comment to the go code to explain what the function does and
launch golangci-lint to make sure it's all good.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>