Add doc for passing arguments to GitOps commands
Add documentation for passing arguments to GitOps commands. We moved all the
documentation in a separate file to make it easier to read and find
since it was becoming too big in the `running.md` file.

Add some unittest as well so to make sure the documentation match the
implementation :upside_down_face:

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>