Fix matching proper PipelineRun with /test
We were always getting the first one of the pipelinerun, i thought the
filtering was already done and it would always be only one PipelineRun
but we re-fetch everything on resolve.Resolve, so redoing the filtering
properly.

Improve e2e to make sure it handles this properly to explicit test the
proper PipelineRun.

Refactor a bit the gitea tests for GitOPS commands to be in its own file.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>