Make sure to check if we can make API calls
We didn't check for the error, which could crash the tekton
controller.

That call is the first API call we make so is the first one that would
crash. Report the run as error if that happen, since it's probably a
GitHub API outage (and a lot of other things would be wrong)

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>