feat(cel): Add unittests for CEL package
This commit refactors the CEL related functions and moves them from
the customparams package to the cel package as the how the folder is
named. The function names have been simplified for clarity. Unit tests
have been added for the refactored functions. The changes also include
updates to the function calls in other packages.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>