feat: Allow non-admin users to get install info
This commit modifies the 'install' command to allow non-admin users to
retrieve installation information for Pipelines-as-Code. Previously, this
command was restricted to admin users only. Now, if the user does not have
rights to list all repos in the cluster, the command will attempt to list
repos in the current namespace instead. The command's help text has been
updated to reflect this change.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>