fix: Correct event-type setting for on-comment prs
This commit fixes the event-type setting for PipelineRuns that match
the `on-comment` annotation. Previously, the `event_type` was set to
`no-ops-comment` instead of `on-comment` when a git provider's issue
comment event matched a comment in an annotation. This fix ensures that
the `event_type` is correctly set.

on-comment is a brand new TP feature, so no worries that we may have
breaking things for users relying on that behaviour.

Issue: https://issue.redhat.com/browse/SRVKP-5779
Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>