Continue when failing the concurrency
When we were failing concurrency we were not continuing to the next one
and the semaphore will not be removing the lock which would end up with
a deadlock.

This only happens when the underlying cluster is becoming really busy
and have issues to update the pipelinerun in progress.

The only way to test this is to modify updatePipelineRunToInProgress.

I have a repo with three pipelinerun test-1, test-2 and test-3 matching
a pull-request. I have a concurrency_limit of 1 in the repo spec.

I run the Pr and the test-1 should run successfully, test-2 should
return an error but here test-3 should get kicked and start.

Here is the patch i have to do to test this, add this to
reconciler/reconciler.go:

```go
func randomError(prn string) error {
  if strings.HasPrefix(prn, "test-2") {
   return fmt.Errorf("DEBUG: 😈 randomly failing this PipelineRun: %s", prn)
  }
  /* if we want to generate more randomness on load testing
  var b [8]byte
  if _, err := rand.Read(b[:]); err != nil {
    return fmt.Errorf("failed to generate random number: %w", err)
  }
  randomNumber := binary.LittleEndian.Uint64(b[:]) % 100

  if randomNumber < 50 {
   return fmt.Errorf("DEBUG: 😈 randomly failing this PipelineRun: %s", prn)
  }

  */
  return nil
}
```

and add at the top of the `updatePipelineRunToInProgress` function:

```go
if err := randomError(pr.GetName()); err != nil {
 return err
}
```

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>